Move build stuff into lib folder & update versions
