+++ .gitignore
@@ -1,4 +1,4 @@
-dependencygraph.json
+lib/dependencygraph.json
 .rpt2_cache
 node_modules
 dist
