+++ lib/asset-plugin.js
