+++ lib/asset-transform-plugin.js
