+++ lib/chunk-name-plugin.js
