+++ lib/consts-plugin.js
