+++ lib/dependency-graph-plugin.js
