+++ lib/ejs-asset-plugin.js
