+++ lib/glsl-plugin.js
