+++ lib/render_static.js
@@ -65,7 +65,7 @@ async function generateShell(file, dependencygraph) {
     dependencygraph,
     icon: findAssetWithName(dependencygraph, "icon-maskable.png").fileName,
     dependencygraph,
-    pkg: require("./package.json"),
+    pkg: require("../package.json"),
     fs,
     title: "PROXX — a game",
     description:
