+++ lib/resource-list-plugin.js
